refactor: restructure Home components under `index` directory
Moved all Home-related components, hooks, and utilities into an `index` subdirectory for improved file organization and consistency. Updated all related imports to align with the new structure.
