+++ apps/nextjs/src/app/(home)/layout.tsx
@@ -1,4 +1,4 @@
-import { Page } from '@/components/pages/home';
+import { Page } from '@/components/pages/home/index';
 
 export default function HomePage() {
   return <Page />;
