+++ apps/nextjs/src/components/pages/home/index/components/content/index.ts
