+++ apps/nextjs/src/components/pages/home/index/components/favorite-projects/index.ts
