+++ apps/nextjs/src/components/pages/home/index/components/projects-container/projects-container.tsx
