+++ apps/nextjs/src/components/pages/home/index/components/recent-projects/recent-projects.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { ProjectsContainer } from '@/components/pages/home/components/projects-container';
+import { ProjectsContainer } from '@/components/pages/home/index/components/projects-container';
 import { useProjectIds } from '@/features/project/store/project';
 
 export const RecentProjects = memo(function RecentProjects() {
