+++ apps/nextjs/src/components/pages/home/index/components/skeleton-home.tsx
