+++ apps/nextjs/src/components/pages/home/index/store/home/type.ts
